+++ src/Pyro4/configuration.py
@@ -18,7 +18,8 @@ class Configuration(object):
                "DETAILED_TRACEBACK", "SOCK_REUSE", "PREFER_IP_VERSION",
                "THREADPOOL_MINTHREADS", "THREADPOOL_MAXTHREADS",
                "THREADPOOL_IDLETIMEOUT", "HMAC_KEY", "AUTOPROXY",
-               "BROADCAST_ADDRS", "NATHOST", "NATPORT", "MAX_MESSAGE_SIZE")
+               "BROADCAST_ADDRS", "NATHOST", "NATPORT", "MAX_MESSAGE_SIZE",
+               "FLAME_ENABLED" )
 
     def __init__(self):
         self.reset()
@@ -51,6 +52,7 @@ class Configuration(object):
         self.AUTOPROXY = True
         self.MAX_MESSAGE_SIZE = 0   # 0 = unlimited
         self.BROADCAST_ADDRS = "<broadcast>, 0.0.0.0"   # comma separated list of broadcast addresses
+        self.FLAME_ENABLED = False
         self.PREFER_IP_VERSION = 4    # 4, 6 or 0 (let OS choose according to RFC 3484)
 
         if useenvironment: